+++ packages/xchain-binance/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Update dependencies
 - Add `getDefaultFees`
 - Add `getSingleAndMultiFees`
+- Add `getDerivePath` helper
 
 # v.4.1.0 (2020-11-20)