+++ packages/xchain-binance/CHANGELOG.md
@@ -1,5 +1,9 @@
 # v.x.x.x (2020-XX-XX)
 
+### Breaking change
+
+- Remove `validateAddress` from `BinanceClient`
+
 # v.4.2.0 (2020-12-11)
 
 ### Update