+++ packages/xchain-binance/CHANGELOG.md
@@ -1,4 +1,4 @@
-# v.x.x.x (2020-XX-XX)
+# v.4.2.1 (2020-12-28)
 
 ### Breaking change